requirements.txt
# Base----------------------------------
albumentations>=1.1.0
matplotlib
opencv-python>=4.1.1.26
opencv-python-headless>=4.1.1.26
Pillow
PyYAML
scikit-image
scikit-learn
scipy
torch>=1.10.0
torchvision>=0.11.0
numpy
protobuf<=3.20.1
pandas

# Extras----------------------------------
pycocotools>=2.0.2
setuptools==59.5.0